rename(components): 상품 컴포넌트 폴더 분리(#114)
category와 product로 나누어 관리하던 상품 목록 관련 컴포넌트를
product/productlist 폴더로 이동하였습니다. 또한 상품 상세 페이지와
관련된 작업은 product/productdetail 폴더를 추가하여 작업할 계획입니다.
